feat: Replace dynamic fee module with custom fee module for Velodrome
