+++ protocols/substreams/unichain-velodrome/build.rs
@@ -8,8 +8,8 @@ fn main() -> Result<(), anyhow::Error> {
     Abigen::new("Pool", "abi/Pool.json")?
         .generate()?
         .write_to_file("src/abi/pool.rs")?;
-    Abigen::new("DynamicSwapFeeModule", "abi/DynamicSwapFeeModule.json")?
+    Abigen::new("CustomSwapFeeModule", "abi/CustomSwapFeeModule.json")?
         .generate()?
-        .write_to_file("src/abi/dynamic_swap_fee_module.rs")?;
+        .write_to_file("src/abi/custom_swap_fee_module.rs")?;
     Ok(())
 }
