+++ protocols/substreams/unichain-velodrome/integration_test.tycho.yaml
@@ -1,7 +1,7 @@
-substreams_yaml_path: ./base-aerodrome-slipstreams.yaml
-protocol_system: "aerodrome_slipstreams"
+substreams_yaml_path: ./unichain-velodrome.yaml
+protocol_system: "velodrome_slipstreams"
 protocol_type_names:
-  - "aerodrome_slipstreams_pool"
+  - "velodrome_slipstreams_pool"
 module_name: "map_protocol_changes"
 skip_balance_check: true
 initialized_accounts:
