+++ protocols/substreams/unichain-velodrome/src/abi/mod.rs
@@ -1,5 +1,5 @@
 #![allow(clippy::all, clippy::pedantic, clippy::nursery)]
 
-pub mod dynamic_swap_fee_module;
+pub mod custom_swap_fee_module;
 pub mod factory;
 pub mod pool;
