+++ protocols/substreams/unichain-velodrome/src/events/burn.rs
@@ -1,6 +1,6 @@
 use crate::{
     abi::pool::events::Burn,
-    pb::tycho::evm::aerodrome::Pool,
+    pb::tycho::evm::velodrome::Pool,
     storage::{constants::TRACKED_SLOTS, pool_storage::SlipstreamsPoolStorage},
 };
 use substreams_ethereum::pb::eth::v2::StorageChange;
@@ -33,9 +33,6 @@ impl EventTrait for Burn {
             pool_storage.get_ticks_changes(vec![&self.tick_upper, &self.tick_lower]);
 
         changed_attributes.extend(changed_ticks);
-
-        let changed_observation = pool_storage.get_all_observations_changes();
-        changed_attributes.extend(changed_observation);
 
         changed_attributes
     }
