+++ protocols/substreams/unichain-velodrome/src/events/collect_fees.rs
@@ -1,7 +1,7 @@
 use super::{BalanceDelta, EventTrait};
 use crate::{
     abi::pool::events::CollectFees,
-    pb::tycho::evm::aerodrome::Pool,
+    pb::tycho::evm::velodrome::Pool,
     storage::{constants::TRACKED_SLOTS, pool_storage::SlipstreamsPoolStorage},
 };
 use substreams_ethereum::pb::eth::v2::StorageChange;
