+++ protocols/substreams/unichain-velodrome/src/events/initialize.rs
@@ -1,10 +1,9 @@
-use substreams::scalar::BigInt;
 use substreams_ethereum::pb::eth::v2::StorageChange;
 
 use super::{BalanceDelta, EventTrait};
 use crate::{
     abi::pool::events::Initialize,
-    pb::tycho::evm::aerodrome::Pool,
+    pb::tycho::evm::velodrome::Pool,
     storage::{constants::TRACKED_SLOTS, pool_storage::SlipstreamsPoolStorage},
 };
 use substreams_helper::storage_change::StorageChangesFilter;
@@ -25,9 +24,8 @@ impl EventTrait for Initialize {
             .collect();
         let pool_storage = SlipstreamsPoolStorage::new(&filtered_storage_changes);
 
-        let mut attributes =
+        let attributes =
             pool_storage.get_changed_attributes(TRACKED_SLOTS.to_vec().iter().collect());
-        attributes.extend(pool_storage.get_observations_changes(vec![&BigInt::zero()]));
 
         attributes
     }
