+++ protocols/substreams/unichain-velodrome/src/modules/1_map_tick_spacing_fee.rs
@@ -1,7 +1,7 @@
 use crate::{
     abi::factory::events::TickSpacingEnabled,
     modules::utils::Params,
-    pb::tycho::evm::aerodrome::{TickSpacingFee, TickSpacingFees},
+    pb::tycho::evm::velodrome::{TickSpacingFee, TickSpacingFees},
 };
 use ethabi::ethereum_types::Address;
 use std::str::FromStr;
