+++ protocols/substreams/unichain-velodrome/src/modules/2_store_tick_spacing_fee.rs
@@ -1,4 +1,4 @@
-use crate::pb::tycho::evm::aerodrome::TickSpacingFees;
+use crate::pb::tycho::evm::velodrome::TickSpacingFees;
 use substreams::store::{StoreNew, StoreSet, StoreSetInt64};
 
 #[substreams::handlers::store]
