+++ protocols/substreams/unichain-velodrome/src/pb/mod.rs
@@ -1,10 +1,10 @@
 // @generated
 pub mod tycho {
     pub mod evm {
-        // @@protoc_insertion_point(attribute:tycho.evm.aerodrome)
-        pub mod aerodrome {
-            include!("tycho.evm.aerodrome.rs");
-            // @@protoc_insertion_point(tycho.evm.aerodrome)
+        // @@protoc_insertion_point(attribute:tycho.evm.velodrome)
+        pub mod velodrome {
+            include!("tycho.evm.velodrome.rs");
+            // @@protoc_insertion_point(tycho.evm.velodrome)
         }
     }
 }
