+++ protocols/substreams/unichain-velodrome/src/storage/constants.rs
@@ -23,34 +23,8 @@ const SQRT_PRICE_X96_SLOT: StorageLocation = StorageLocation {
 const CURRENT_TICK_SLOT: StorageLocation =
     StorageLocation { name: "tick", slot: SLOT0, offset: 20, number_of_bytes: 3, signed: true };
 
-const OBSERVATION_INDEX_SLOT: StorageLocation = StorageLocation {
-    name: "observationIndex",
-    slot: SLOT0,
-    offset: 23,
-    number_of_bytes: 2,
-    signed: false,
-};
-
-const OBSERVATION_CARDINALITY_SLOT: StorageLocation = StorageLocation {
-    name: "observationCardinality",
-    slot: SLOT0,
-    offset: 25,
-    number_of_bytes: 2,
-    signed: false,
-};
-
 pub(crate) const TICKS_MAP_SLOT: [u8; 32] =
     hex!("0000000000000000000000000000000000000000000000000000000000000011");
 
-pub(crate) const OBSERVATIONS: [u8; 32] =
-    hex!("0000000000000000000000000000000000000000000000000000000000000014");
-
-pub(crate) const OBSERVATIONS_BASE_U64: u64 = 20;
-
-pub(crate) const TRACKED_SLOTS: [StorageLocation; 5] = [
-    LIQUIDITY_SLOT,
-    SQRT_PRICE_X96_SLOT,
-    CURRENT_TICK_SLOT,
-    OBSERVATION_INDEX_SLOT,
-    OBSERVATION_CARDINALITY_SLOT,
-];
+pub(crate) const TRACKED_SLOTS: [StorageLocation; 3] =
+    [LIQUIDITY_SLOT, SQRT_PRICE_X96_SLOT, CURRENT_TICK_SLOT];
