+++ protocols/substreams/unichain-velodrome/unichain-velodrome.yaml
@@ -1,14 +1,14 @@
 specVersion: v0.1.0
 package:
-  name: "base_aerodrome_slipstreams"
+  name: "unichain_velodrome"
   version: v0.1.0
-  url: "https://github.com/propeller-heads/tycho-protocol-sdk/tree/main/protocols/substreams/base-aerodrome-slipstreams"
+  url: "https://github.com/propeller-heads/tycho-protocol-sdk/tree/main/substreams/unichain-velodrome"
 
 protobuf:
   files:
-    - tycho/protocols/adapter-integration/evm/v1/common.proto
-    - tycho/protocols/adapter-integration/evm/v1/entity.proto
-    - tycho/protocols/adapter-integration/evm/v1/utils.proto
+    - tycho/evm/v1/common.proto
+    - tycho/evm/v1/entity.proto
+    - tycho/evm/v1/utils.proto
     - slipstreams.proto
   importPaths:
     - ./proto/v1
@@ -17,21 +17,21 @@ protobuf:
 binaries:
   default:
     type: wasm/rust-v1
-    file: ../target/wasm32-unknown-unknown/release/base_aerodrome_slipstreams.wasm
+    file: ../target/wasm32-unknown-unknown/release/unichain_velodrome.wasm
 
 modules:
   - name: map_tick_spacing_fee
     kind: map
-    initialBlock: 13843704
+    initialBlock: 9386540
     inputs:
       - params: string
       - source: sf.ethereum.type.v2.Block
     output:
-      type: proto:tycho.evm.aerodrome.TickSpacingFees
+      type: proto:tycho.evm.velodrome.TickSpacingFees
 
   - name: store_tick_spacing_fee
     kind: store
-    initialBlock: 13843704
+    initialBlock: 9386540
     updatePolicy: set
     valueType: proto:slipstreams.TickSpacingFees
     inputs:
@@ -39,7 +39,7 @@ modules:
 
   - name: map_pools_created
     kind: map
-    initialBlock: 13843704
+    initialBlock: 9386540
     inputs:
       - params: string
       - source: sf.ethereum.type.v2.Block
@@ -49,7 +49,7 @@ modules:
 
   - name: store_pools
     kind: store
-    initialBlock: 13843704
+    initialBlock: 9386540
     updatePolicy: set_if_not_exists
     valueType: proto:slipstreams.Pool
     inputs:
@@ -57,7 +57,7 @@ modules:
 
   - name: map_balance_changes
     kind: map
-    initialBlock: 13843704
+    initialBlock: 9386540
     inputs:
       - source: sf.ethereum.type.v2.Block
       - store: store_pools
@@ -66,7 +66,7 @@ modules:
 
   - name: store_pools_balances
     kind: store
-    initialBlock: 13843704
+    initialBlock: 9386540
     updatePolicy: add
     valueType: bigint
     inputs:
@@ -74,7 +74,7 @@ modules:
 
   - name: map_protocol_changes
     kind: map
-    initialBlock: 13843704
+    initialBlock: 9386540
     inputs:
       - params: string
       - source: sf.ethereum.type.v2.Block
@@ -87,6 +87,6 @@ modules:
       type: proto:tycho.evm.v1.BlockChanges
 
 params:
-  map_tick_spacing_fee: "factories[]=5e7BB104d84c7CB9B682AaC2F3d509f5F406809A&factories[]=aDe65c38CD4849aDBA595a4323a8C7DdfE89716a&dynamic_fee_modules[]=DB45818A6db280ecfeB33cbeBd445423d0216b5D&dynamic_fee_modules[]=00cB12a1c84dfC1b9c70734C0385E769Bc86e9Ef"
-  map_pools_created: "factories[]=5e7BB104d84c7CB9B682AaC2F3d509f5F406809A&factories[]=aDe65c38CD4849aDBA595a4323a8C7DdfE89716a&dynamic_fee_modules[]=DB45818A6db280ecfeB33cbeBd445423d0216b5D&dynamic_fee_modules[]=00cB12a1c84dfC1b9c70734C0385E769Bc86e9Ef"
-  map_protocol_changes: "factories[]=5e7BB104d84c7CB9B682AaC2F3d509f5F406809A&factories[]=aDe65c38CD4849aDBA595a4323a8C7DdfE89716a&dynamic_fee_modules[]=DB45818A6db280ecfeB33cbeBd445423d0216b5D&dynamic_fee_modules[]=00cB12a1c84dfC1b9c70734C0385E769Bc86e9Ef"
+  map_tick_spacing_fee: "factories[]=04625b046c69577efc40e6c0bb83cdbafab5a55f&custom_fee_modules[]=6812eefC19deB79D5191b52f4B763260d9F3C238"
+  map_pools_created: "factories[]=04625b046c69577efc40e6c0bb83cdbafab5a55f&custom_fee_modules[]=6812eefC19deB79D5191b52f4B763260d9F3C238"
+  map_protocol_changes: "factories[]=04625b046c69577efc40e6c0bb83cdbafab5a55f&custom_fee_modules[]=6812eefC19deB79D5191b52f4B763260d9F3C238"
